@@ -1,70 +1,23 @@
 package top.yangyl.datasource.test;
 
-import com.zaxxer.hikari.HikariConfig;
-import com.zaxxer.hikari.HikariDataSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.core.io.support.PropertiesLoaderUtils;
 
 import javax.sql.DataSource;
-import java.io.IOException;
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
+import java.sql.*;
+import java.util.ArrayList;
+import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Properties;
+import java.util.Map;
 
 public class DBUtils {
 
     private Logger logger= LoggerFactory.getLogger(DBUtils.class);
 
-    private static String dbUrl;
-    private static String dbUser;
-    private static String dbPw;
-    private static DataSource dataSource;
-    private static String driverClassName;
-
-    static {
-        try {
-            Properties properties = PropertiesLoaderUtils.loadProperties(new ClassPathResource("jdbc.properties"));
-            dbUrl=properties.getProperty("jdbc.url");
-            dbUser=properties.getProperty("jdbc.username");
-            dbPw=properties.getProperty("jdbc.password");
-            driverClassName=properties.getProperty("jdbc.driver");
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    public static DataSource getDataSource(){
-        synchronized (DBUtils.class){
-            if(dataSource==null){
-                initDataSource();
-            }
-            return dataSource;
-        }
-    }
-
-    private static void initDataSource(){
-        HikariConfig config=new HikariConfig();
-        config.setDriverClassName(driverClassName);
-        config.setJdbcUrl(dbUrl);
-        config.setUsername(dbUser);
-        config.setPassword(dbPw);
-        config.setAutoCommit(true);
-        config.setTransactionIsolation("2");
-        config.setMinimumIdle(10);
-        config.setMaximumPoolSize(20);
-        dataSource=new HikariDataSource(config);
-    }
-
-    public static int execute(String sql, List<Object> params) {
+    public static int execute(DataSource ds,String sql, List<Object> params) {
         Connection connection = null;
         PreparedStatement preparedStatement = null;
         ResultSet resultSet = null;
-        DataSource ds=getDataSource();
         try {
             connection = ds.getConnection();
             connection.setAutoCommit(false);
@@ -85,6 +38,41 @@ public class DBUtils {
             } catch (SQLException e1) {
                 e1.printStackTrace();
             }
+            throw new RuntimeException(e);
+        } finally {
+            Closer.closeQuietly(resultSet, preparedStatement, connection);
+        }
+    }
+
+
+    public static List<Map<String, Object>> query(DataSource ds, String sql, List<Object> params) {
+        Connection connection = null;
+        PreparedStatement preparedStatement = null;
+        ResultSet resultSet = null;
+        try {
+            connection = ds.getConnection();
+            preparedStatement = connection.prepareStatement(sql);
+            if (params != null && params.size() > 0) {
+                int size = params.size();
+                for (int i = 0; i < size; i++) {
+                    preparedStatement.setObject(i + 1, params.get(i));
+                }
+            }
+            resultSet = preparedStatement.executeQuery();
+            List<Map<String, Object>> results = new ArrayList<>();
+            while (resultSet.next()) {
+                Map<String, Object> map = new LinkedHashMap<>();
+                ResultSetMetaData metaData = resultSet.getMetaData();
+                int columnCount = metaData.getColumnCount();
+                for (int i = 0; i < columnCount; i++) {
+                    String columnLabel = metaData.getColumnLabel(i + 1);
+                    map.put(columnLabel, resultSet.getObject(columnLabel));
+                }
+                results.add(map);
+            }
+
+            return results;
+        } catch (SQLException e) {
             throw new RuntimeException(e);
         } finally {
             Closer.closeQuietly(resultSet, preparedStatement, connection);
